pyproject.toml
[tool.poetry]
name = "btu-virtual-university"
version = "0.1.0"
description = "BTU Virtual University – Multi-Agentic AI Framework"
authors = ["BTU Engineering Team"]
readme = "README.md"
packages = [{include = "btu"}]

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.40.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
faiss-cpu = "^1.7.0"
sentence-transformers = "^3.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.3.0"
structlog = "^24.2.0"
httpx = "^0.27.0"
anyio = "^4.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-httpx = "^0.30.0"
coverage = "^7.5.0"
ruff = "^0.4.0"
mypy = "^1.10.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
